Preserve Artistic Composition
Shifts the center of crop to best represent the original composition of
the image. This is done by weighing the location of hotspots relative to
the original image and shifting to cropping center to re-frame the
hotspots to roughly match the original composition.

Example:
If a single hot spot is about 25% form the left and 30% from the top,
the cropped images will have that hotspot re-framed approximately
25% from the left and 30% form the top, instead of just centered.